+++ src/app/pages/companies/companies.component.ts
@@ -12,8 +12,8 @@ import { RippleModule } from 'primeng/ripple';
 import { TableModule } from 'primeng/table';
 import { TagModule } from 'primeng/tag';
 import { ToastModule } from 'primeng/toast';
-import { CompanyReqDto } from '../../models/dto/company/company.req.dto';
-import { CompanyResDto } from '../../models/dto/company/company.res.dto';
+import { CompanyReqDto } from '../../dto/company/company.req.dto';
+import { CompanyResDto } from '../../dto/company/company.res.dto';
 
 @Component({
   selector: 'app-companies',
@@ -94,20 +94,20 @@ export class CompaniesComponent {
   }
 
   onRowEditSave(company: CompanyResDto) {
-    const editedCompany: CompanyReqDto = this.companyForm.getRawValue() as CompanyReqDto;
+    // const editedCompany: CompanyReqDto = this.companyForm.getRawValue() as CompanyReqDto;
 
-    // Update the original data with the edited data
-    const index = this.companies.findIndex(c => c.id === company.id);
-    if (index !== -1) {
-      this.companies[index] = { ...this.companies[index], ...editedCompany };
-    }
+    // // Update the original data with the edited data
+    // const index = this.companies.findIndex(c => c.id === company.id);
+    // if (index !== -1) {
+    //   this.companies[index] = { ...this.companies[index], ...editedCompany };
+    // }
 
-    // Clear the form group
-    this.companyForm.reset();
-    this.isEditing = false;
-    this.originalCompanyData = null;
+    // // Clear the form group
+    // this.companyForm.reset();
+    // this.isEditing = false;
+    // this.originalCompanyData = null;
 
-    this.messageService.add({ severity: 'success', summary: 'Success', detail: `User is updated` });
+    // this.messageService.add({ severity: 'success', summary: 'Success', detail: `User is updated` });
   }
 
   onRowEditCancel(company: CompanyResDto, index: number) {
