+++ src/app/services/payroll.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from "@angular/core";
-import { PayrollsResDto } from "../models/dto/payrolls/payrolls.res.dto";
 import { BaseService } from "./base/base.service";
+import { PayrollsResDto } from "../dto/payrolls/payrolls.res.dto";
 
 @Injectable({
   providedIn: "root"
