+++ src/app/services/user/user.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import { BaseService } from '../base/base.service';
-import { UserResDto } from '../../models/dto/user/user.res.dto';
-import { UserReqDto } from '../../models/dto/user/user.req.dto';
-import { InsertResDto } from '../../models/dto/insert.res.dto';
+import { UserResDto } from '../../dto/user/user.res.dto';
+import { InsertResDto } from '../../dto/insert.res.dto';
+import { UserReqDto } from '../../dto/user/user.req.dto';
 
 @Injectable({
   providedIn: 'root'
